+++ test/queues/int.spec.ts
@@ -4,9 +4,7 @@ import {
   _internal_resetJobSystemGlobals,
   type JobTaskStatus,
   type Payload,
-  type SanitizedConfig,
 } from 'payload'
-import { getPayload, migrateCLI } from 'payload'
 import { wait } from 'payload/shared'
 import { fileURLToPath } from 'url'
 
@@ -21,7 +19,7 @@ const { email, password } = devUser
 const filename = fileURLToPath(import.meta.url)
 const dirname = path.dirname(filename)
 
-describe('Queues', () => {
+describe('Queues - Payload', () => {
   let payload: Payload
   let restClient: NextRESTClient
   let token: string
@@ -1465,74 +1463,5 @@ describe('Queues', () => {
 
     expect(allSimples.totalDocs).toBe(1)
     expect(allSimples?.docs?.[0]?.title).toBe('hello!')
-  })
-})
-
-describe('Queues - CLI', () => {
-  let config: SanitizedConfig
-  beforeAll(async () => {
-    ;({ config } = await initPayloadInt(dirname, undefined, false))
-  })
-
-  it('ensure consecutive getPayload call with cron: true will autorun jobs', async () => {
-    const payload = await getPayload({
-      config,
-    })
-
-    await payload.jobs.queue({
-      workflow: 'inlineTaskTest',
-      queue: 'autorunSecond',
-      input: {
-        message: 'hello!',
-      },
-    })
-
-    process.env.PAYLOAD_DROP_DATABASE = 'false'
-
-    // Second instance of payload with the only purpose of running cron jobs
-    const _payload2 = await getPayload({
-      config,
-      cron: true,
-    })
-
-    await waitUntilAutorunIsDone({
-      payload,
-      queue: 'autorunSecond',
-    })
-
-    const allSimples = await payload.find({
-      collection: 'simple',
-      limit: 100,
-    })
-
-    expect(allSimples.totalDocs).toBe(1)
-    expect(allSimples?.docs?.[0]?.title).toBe('hello!')
-
-    // Shut down safely:
-    // Ensure no new crons are scheduled
-    _internal_jobSystemGlobals.shouldAutoRun = false
-    _internal_jobSystemGlobals.shouldAutoSchedule = false
-    // Wait 3 seconds to ensure all currently-running crons are done. If we shut down the db while a function is running, it can cause issues
-    // Cron function runs may persist after a test has finished
-    await wait(3000)
-    // Now we can destroy the payload instance
-    await _payload2.destroy()
-    await payload.destroy()
-    _internal_resetJobSystemGlobals()
-  })
-
-  it('can run migrate CLI without jobs attempting to run', async () => {
-    await migrateCLI({
-      config,
-      parsedArgs: {
-        _: ['migrate'],
-      },
-    })
-
-    // Wait 3 seconds to let potential autorun crons trigger
-    await new Promise((resolve) => setTimeout(resolve, 3000))
-
-    // Expect no errors. Previously, this would throw an "error: relation "payload_jobs" does not exist" error
-    expect(true).toBe(true)
   })
 })
